+++ front/src/app/modules/utilisateurs/components/pages/utilisateur-tirage-page.component.ts
@@ -1,9 +1,10 @@
 import { HttpClient } from '@angular/common/http';
 import { Component } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { Observable } from 'rxjs';
-import { map, switchMap } from 'rxjs/operators';
-import { IUtilisateur, StatutTirage } from '../../../../../../../shared/domaine';
+import { Observable, combineLatest } from 'rxjs';
+import { map, switchMap, filter } from 'rxjs/operators';
+
+import { StatutTirage } from '../../../../../../../shared/domaine';
 import { TirageDTO } from '../../../../../../../back/src/utilisateurs/dto/tirage.dto';
 import { environment } from '../../../../../environments/environment';
 import { TiragesService } from '../../../tirages/services/tirages.service';
@@ -14,27 +15,31 @@ import { TiragesService } from '../../../tirages/services/tirages.service';
   styleUrls: ['./utilisateur-tirage-page.component.scss']
 })
 export class UtilisateurTiragePageComponent {
-  idUtilisateur$: Observable<IUtilisateur['id']>;
-  tirage$: Observable<TirageDTO>;
-  estTirageCree$: Observable<boolean>;
+  params$: Observable<{
+    idUtilisateur: number,
+    idTirage: number
+  }>;
+  tirage$: Observable<TirageDTO & { lance: boolean }>;
 
   constructor(
     http: HttpClient,
     route: ActivatedRoute
   ) {
-    this.idUtilisateur$ = route.params.pipe(
-      map(p => +p.idUtilisateur)
+    this.params$ = route.paramMap.pipe(
+      map(pm => ({
+        idUtilisateur: parseInt(pm.get('idUtilisateur')),
+        idTirage: parseInt(pm.get('idTirage'))
+      })),
+      filter(params => !isNaN(params.idUtilisateur) && !isNaN(params.idTirage))
     );
-    this.tirage$ = route.params.pipe(
-      map(p => [p.idUtilisateur, p.idTirage]),
-        switchMap(([idUtilisateur, idTirage]) => {
-          return http.get<TirageDTO>(environment.backUrl + `/utilisateurs/${idUtilisateur}/tirages/${idTirage}`)
-        }
-      ),
-      map(TiragesService.formateDates())
+    this.tirage$ = this.params$.pipe(
+      switchMap(({ idUtilisateur, idTirage }) => {
+        return http.get<TirageDTO>(environment.backUrl + `/utilisateurs/${idUtilisateur}/tirages/${idTirage}`)
+      }),
+      map(TiragesService.formateDates()),
+      map(tirage => Object.assign(tirage, {
+        lance: tirage.statut !== StatutTirage.CREE
+      }))
     );
-    this.estTirageCree$ = this.tirage$.pipe(
-      map(t => t.statut === StatutTirage.CREE)
-    )
   }
 }